fix issu #685
